:recycle: Fix login middleware file after code review
